fix: handle maps with `__struct__` key
Fixes  #57

Ecto.Schema has a typespec that looks like this:

```elixir
 schema() :: %{
  optional(atom()) => any(),
  :__struct__ => atom(),
  :__meta__ => Ecto.Schema.Metadata.t()
}
```

Because of the `__struct__` key, it looks like a struct, but really isn't and that was causing a `CaseClauseError` when trying to type-check an Ecto.Changeset.t() (which has a `schema()` inside of it.

This change adds a case clause that matches this case and then recursively ensures that the value of the `__struct__` key matches the specified type and that the rest of the fields also match the rest of the type.